Remove tracking from useParty
