+++ src/utils/tracking.ts
@@ -1,4 +1,7 @@
+import { useContext, useEffect } from 'react'
+import { useParams } from 'react-router-dom'
 import { v4 } from 'uuid'
+import { PlayerContext } from '../AuthContext'
 
 export const track = (event: string, args?: {}) => {
     window.heap?.track(event, args)
@@ -21,4 +24,33 @@ export const useStaticUID = () => {
         return uid
     }
     return storedUID
+}
+
+const staticUID = useStaticUID()
+
+export const useGlobalTracking = () => {
+    const { player } = useContext(PlayerContext)
+    const { partyId } = useParams()
+
+    // Init trackers
+    useEffect(() => {
+        if (partyId && window.$crisp && player) {
+            window.$crisp.push([
+                'set',
+                'user:nickname',
+                [`user:${player.username || player.userId.toString()}`],
+            ])
+            window.$crisp.push([
+                'set',
+                'session:data',
+                [[['partyId', partyId]]],
+            ])
+            window.heap?.identify(staticUID)
+            window.heap?.addUserProperties({
+                partyId,
+                partyUserId: player.userId,
+                ...(player.username ? { username: player.username } : {}),
+            })
+        }
+    }, [partyId, player])
 }
